+++ wetterdienst/api.py
@@ -10,6 +10,7 @@ from wetterdienst.provider.dwd.observation import DwdObservationRequest
 from wetterdienst.provider.dwd.radar import DwdRadarValues
 from wetterdienst.provider.eccc.observation import EcccObservationRequest
 from wetterdienst.provider.noaa.ghcn.api import NoaaGhcnRequest
+from wetterdienst.provider.wsv.pegel.api import WsvPegelRequest
 from wetterdienst.util.enumeration import parse_enumeration_from_template
 from wetterdienst.util.parameter import DatasetTreeCore
 
@@ -25,6 +26,9 @@ class ApiEndpoints(DatasetTreeCore):
 
     class NOAA(Enum):
         GHCN = NoaaGhcnRequest
+
+    class WSV(Enum):
+        PEGEL = WsvPegelRequest
 
 
 class Wetterdienst: