+++ wetterdienst/provider/wsv/pegel/__init__.py
@@ -1,0 +1,1 @@
+from wetterdienst.provider.wsv.pegel.api import WsvPegelParameter, WsvPegelRequest